ref: allow custom message on input, moved functions to separate files
