+++ basic_operations.py
@@ -1,0 +1,17 @@
+def add(a, b):
+    return a + b
+
+def subtract(a, b):
+    return a - b
+
+def multiply(a, b):
+    return a * b
+
+def divide(a, b):
+    if b != 0:
+        return a / b
+    else:
+        return "Помилка: ділення на нуль"
+
+def modulo(a, b):
+    return a % b
