+++ sin_cos_operations.py
@@ -1,9 +1,9 @@
 import math
 
-
+# OLGA: sin(a)
 def sin_number(a: int) -> int:
     return math.sin(a)
 
-
+# OLGA: cos(a)
 def cos_number(a: int) -> int:
     return math.cos(a)
